完善Shell布局
