+++ MapBoard.MAUI/Mapping/MapViewHelper.cs
@@ -3,53 +3,17 @@ using Esri.ArcGISRuntime.Geometry;
 using Esri.ArcGISRuntime.Mapping;
 using Esri.ArcGISRuntime.Rasters;
 using Esri.ArcGISRuntime.Maui;
-using MapBoard.Mapping;
 using MapBoard.Model;
 using System.ComponentModel;
-using System.Drawing;
-using System.Runtime.InteropServices;
 using FzLib;
-using Newtonsoft.Json.Linq;
-using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Drawing.Imaging;
-using System.IO;
-using System.Linq;
-using System.Reflection;
 using System.Text.RegularExpressions;
-using System.Threading.Tasks;
-using System.Windows;
 using Map = Esri.ArcGISRuntime.Mapping.Map;
 using Esri.ArcGISRuntime.UI;
 using static MapBoard.Util.GeometryUtility;
 
-namespace MapBoard
+namespace MapBoard.Mapping
 {
-    public partial class MainPage : ContentPage
-    {
-        int count = 0;
-
-        public MainPage()
-        {
-            InitializeComponent();
-            map.Map = new Esri.ArcGISRuntime.Mapping.Map();
-
-            map.Map.Basemap = new Esri.ArcGISRuntime.Mapping.Basemap(new WebTiledLayer(""));
-            map.Margin = new Thickness(-72);
-            map.LoadAsync();
-        }
-
-        private void OnCounterClicked(object sender, EventArgs e)
-        {
-
-
-            //SemanticScreenReader.Announce(CounterBtn.Text);
-        }
-
-
-    }
-    public static class GeoViewHelper
+    public static class MapViewHelper
     {
         public static List<BaseLayerInfo> BaseLayers { get; set; } = new List<BaseLayerInfo>()
         {
@@ -430,9 +394,9 @@ namespace MapBoard
             {
                 if (geometry.SpatialReference.Wkid != SpatialReferences.WebMercator.Wkid)
                 {
-                    geometry = GeometryEngine.Project(geometry, SpatialReferences.WebMercator);
+                    geometry = geometry.Project(SpatialReferences.WebMercator);
                 }
-                geometry = GeometryEngine.Buffer(geometry, 500);
+                geometry = geometry.Buffer(500);
             }
             var extent = geometry.Extent;
             if (double.IsNaN(extent.Width) || double.IsNaN(extent.Height) || extent.Width == 0 || extent.Height == 0)
@@ -467,7 +431,7 @@ namespace MapBoard
                     var point1 = new MapPoint(envelope.XMin, envelope.YMin);
                     var point2 = new MapPoint(envelope.XMax, envelope.YMax);
                     envelope = new Envelope(point1.RegularizeWebMercatorPoint(), point2.RegularizeWebMercatorPoint());
-                    await ZoomToGeometryAsync(mapView, envelope, false);
+                    await mapView.ZoomToGeometryAsync(envelope, false);
                 }
                 catch
                 {
