+++ MapBoard.MAUI/Pages/LayerPage.xaml.cs
@@ -1,0 +1,9 @@
+namespace MapBoard.Pages;
+
+public partial class LayerPage : ContentPage
+{
+	public LayerPage()
+	{
+		InitializeComponent();
+	}
+}
