+++ MapBoard.MAUI/Views/LayerListView.xaml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<ContentView xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             x:Class="MapBoard.Views.LayerListView">
+    <ListView>
+</ListView>
+</ContentView>
