+++ MapBoard.MAUI/Views/LayerListView.xaml.cs
@@ -1,0 +1,9 @@
+namespace MapBoard.Views;
+
+public partial class LayerListView : ContentView
+{
+	public LayerListView()
+	{
+		InitializeComponent();
+	}
+}
